% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gabor_kernel.R
\name{calculate_gabor_params}
\alias{calculate_gabor_params}
\title{Calculate Gabor Parameters}
\usage{
calculate_gabor_params()
}
\value{
A list containing:
- \code{gamma}: Scale factor for the Gaussian envelope in x.
- \code{eta}: Scale factor for the Gaussian envelope in y.
}
\description{
Computes the parameters for the Gabor kernel based on predefined constants
from the reference paper.
}
