% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property.R
\name{convert_from_to}
\alias{convert_from_to}
\title{Convert value from one wave property to another via the shortest path
in the graph of all property types.}
\usage{
convert_from_to(x, start_node, end_node)
}
\description{
Convert value from one wave property to another via the shortest path
in the graph of all property types.
}
