% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{convolution_with_kernel}
\alias{convolution_with_kernel}
\title{Convolution with Kernel}
\usage{
convolution_with_kernel(complex_matrix, kernel)
}
\arguments{
\item{complex_matrix}{The matrix to convolve with}

\item{kernel}{The kernel}
}
\value{
A convolved matrix
}
\description{
Applies a kernel to a complex matrix
}
