% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gabor_kernel.R
\name{create_gabor_kernel}
\alias{create_gabor_kernel}
\title{Create Gabor Kernel}
\usage{
create_gabor_kernel(kernel_size, gamma, eta, orientation, f, psi)
}
\arguments{
\item{kernel_size}{Size of the kernel (must be odd).}

\item{gamma}{Scale factor for the Gaussian envelope in x.}

\item{eta}{Scale factor for the Gaussian envelope in y.}

\item{orientation}{Orientation of the filter in radians.}

\item{f}{Frequency of the sinusoidal wave.}

\item{psi}{Phase offset of the sinusoidal wave.}
}
\value{
A 2D matrix representing the Gabor kernel.
}
\description{
Generates a Gabor kernel with specific parameters.
}
