% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_rds}
\alias{data_rds}
\title{A utility function to cache RDS files in the project's data directory}
\usage{
data_rds(code, filename)
}
\arguments{
\item{code}{The code to execute if the RDS file does not exist.}

\item{filename}{The filename for the RDS file (without path).}
}
\value{
The result of either loading the RDS file or executing the code.
}
\description{
This function checks for the existence of an RDS file in the project's
data directory. If it exists, the RDS file is loaded. If not, the
function passed to \code{code} is executed, and the result is saved to
an RDS file in the data directory.
}
