% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{.spectrum}
\alias{.spectrum}
\title{Internal spectrum constructor with validation and component combination}
\usage{
.spectrum(component, amplitude)
}
\description{
Internal spectrum constructor with validation and component combination
}
\keyword{internal}
