% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property_graph.R
\name{filter_graph_by}
\alias{filter_graph_by}
\title{Filter the graph of wave properties by path length and edge types}
\usage{
filter_graph_by(path_length, relationships = NULL)
}
\arguments{
\item{path_length}{An integer specifying the number of edges (hops) between nodes.}

\item{relationships}{A vector of relationships to filter edges by. If NULL, all relationships are considered.}
}
\value{
A data frame with pairs of nodes separated by the specified path length.
}
\description{
Filter the graph of wave properties by path length and edge types
}
