% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_heisen_spectrum.R
\name{frequency_heisen_spectrum}
\alias{frequency_heisen_spectrum}
\title{Create a Frequency Heisen Spectrum}
\usage{
frequency_heisen_spectrum(
  frequency,
  amplitude = NULL,
  reference_component = NULL
)
}
\arguments{
\item{frequency}{Either a numeric vector of frequencies or a list with named \code{frequency} and \code{amplitude} vectors.}

\item{amplitude}{A numeric vector of amplitudes, if \code{frequency} is a numeric vector.}

\item{reference_component}{Used to compute the fundamental frequency from the cycle length}
}
\value{
An object of class \code{frequency_heisen_spectrum} that inherits from \code{spectrum}.
}
\description{
Extends the \code{frequency_spectrum} class specifically for frequency spectra.
Accepts either direct numeric vectors or a list containing \code{frequency} and \code{amplitude}.
}
