% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_spectrum.R
\name{frequency_spectrum}
\alias{frequency_spectrum}
\title{Create a frequency spectrum}
\usage{
frequency_spectrum(frequency, amplitude)
}
\arguments{
\item{frequency}{A numeric vector of frequencies}

\item{amplitude}{A numeric vector of amplitudes}
}
\value{
An object of class "frequency_spectrum" that inherits from "spectrum"
}
\description{
Create a frequency spectrum
}
