% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_media.R
\name{idealized_spatial_frequency_map}
\alias{idealized_spatial_frequency_map}
\title{Spatial Frequency Map}
\usage{
idealized_spatial_frequency_map(nrows, ncols)
}
\arguments{
\item{nrows}{Number of rows in the image.}

\item{ncols}{Number of columns in the image.}
}
\value{
A matrix with each cell containing spatial frequency values as a named vector.
}
\description{
Creates a matrix where each cell contains a named vector of spatial frequencies (x, y).
}
