% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_sbg.R
\name{kernel_sbg}
\alias{kernel_sbg}
\title{Create SBG Kernel}
\usage{
kernel_sbg(kernel_size)
}
\arguments{
\item{kernel_size}{Size of the kernel (must be odd).}
}
\value{
A 2D matrix representing the SBG kernel.
}
\description{
Generates a SBG kernel with specific parameters.
}
