% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lcm_integers}
\alias{lcm_integers}
\title{Least Common Multiple of Integer Vector}
\usage{
lcm_integers(x)
}
\arguments{
\item{x}{A numeric vector of integers.}
}
\value{
A single numeric value representing the least common multiple.
}
\description{
Computes the least common multiple of a vector of integers.
}
