% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_wavenumber_2D_spectrum.R
\name{linear_wavenumber_2D_spectrum}
\alias{linear_wavenumber_2D_spectrum}
\title{Constructor for linear_wavenumber_2D_spectrum}
\usage{
linear_wavenumber_2D_spectrum(spectrum)
}
\arguments{
\item{spectrum}{A matrix containing the output of a 2D FFT.}
}
\value{
An object of class 'linear_wavenumber_2D_spectrum'.
}
\description{
This function creates an S3 object representing a 2D Fourier Transform spectrum
with linear wavenumbers, storing the real and imaginary parts as sparse Matrices.
}
