% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_heisen_spectrum.R
\name{plot.frequency_heisen_spectrum}
\alias{plot.frequency_heisen_spectrum}
\title{Plot a frequency spectrum with spikes using ggplot2 and theme_homey}
\usage{
\method{plot}{frequency_heisen_spectrum}(x, rectangles = numeric(0), title = NULL, ...)
}
\arguments{
\item{x}{An object of class "frequency_heisen_spectrum".}

\item{rectangles}{Optional: A numeric vector specifying positions for additional rectangles.}

\item{title}{An optional character string for the plot title.}
}
\description{
Provides setup for frequency-specific details before calling the core plot function.
}
