% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_spectrum.R
\name{plot.frequency_spectrum}
\alias{plot.frequency_spectrum}
\title{Plot method for frequency_spectrum}
\usage{
\method{plot}{frequency_spectrum}(x, y = NULL, ...)
}
\arguments{
\item{x}{An object of class "frequency_spectrum"}

\item{y}{Unused, for compatibility with the generic plot function}

\item{...}{Additional graphical parameters}
}
\description{
Plots the frequency and amplitude of a frequency_spectrum object.
}
