% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_waveform.R
\name{plot.linear_waveform}
\alias{plot.linear_waveform}
\title{Plot method for linear_waveform}
\usage{
\method{plot}{linear_waveform}(
  x,
  label = "",
  space_time_range = 25,
  resolution = 300,
  line_plot_resolution = 1000,
  beat_wavelength_spectrum_color = colors_homey$beat,
  ...
)
}
\arguments{
\item{x}{A linear_waveform object.}

\item{label}{A label for the plot.}

\item{space_time_range}{The range of space and time for plotting.}

\item{resolution}{Resolution of the 2D plot.}

\item{line_plot_resolution}{Resolution of the line plots.}

\item{beat_wavelength_spectrum_color}{Color for the beat wavelength spectrum overlay.}

\item{...}{Additional parameters passed to the superclass plot method.}
}
\description{
Overrides the plot function for linear_waveform to include beat_wavelength_spectrum as an overlay.
Calls the superclass plot method, while passing the beat_wavelength_spectrum to be displayed on top of the base wavelength spectrum.
}
