% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{plot.signal}
\alias{plot.signal}
\title{Plot method for signal objects}
\usage{
\method{plot}{signal}(x, label = "", coordinate_range = c(0, 1), resolution = 300)
}
\arguments{
\item{x}{A signal object.}

\item{label}{An optional label for the plot.}

\item{coordinate_range}{A numeric vector of length 2 specifying the start and stop values for the coordinate range.}

\item{resolution}{Number of points to sample within the range.}
}
\description{
Creates a plot of the signal's amplitude over a specified coordinate range.
}
