% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superposed_wave.R
\name{plot.superposed_wave}
\alias{plot.superposed_wave}
\title{Plot method for superposed_wave}
\usage{
\method{plot}{superposed_wave}(
  x,
  label = "",
  space_time_range = 25,
  resolution = 99,
  line_plot_resolution = 99,
  beat_wavelength_spectrum_color = colors_homey$beat,
  ...
)
}
\arguments{
\item{x}{A superposed_wave object.}

\item{label}{A label for the plot.}

\item{space_time_range}{The range of space and time for plotting.}

\item{resolution}{Resolution of the 2D plot.}

\item{line_plot_resolution}{Resolution of the line plots.}

\item{beat_wavelength_spectrum_color}{Color for the beat wavelength spectrum overlay.}

\item{...}{Additional parameters passed to the superclass plot method.}
}
\description{
Overrides the plot function for superposed_wave to include beat_wavelength_spectrum as an overlay.
Calls the superclass plot method, while passing the beat_wavelength_spectrum to be displayed on top of the base wavelength spectrum.
}
