% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelength_spectrum.R
\name{plot.wavelength_spectrum}
\alias{plot.wavelength_spectrum}
\title{Plot a wavelength spectrum with spikes using ggplot2 and theme_homey}
\usage{
\method{plot}{wavelength_spectrum}(x, rectangles = numeric(0), title = NULL, ...)
}
\arguments{
\item{x}{An object of class "wavelength_spectrum".}

\item{rectangles}{Optional: A numeric vector specifying positions for additional rectangles.}

\item{title}{An optional character string for the plot title.}
}
\description{
Provides setup for wavelength-specific details before calling the core plot function.
}
