% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{print.signal}
\alias{print.signal}
\title{Print method for signal objects}
\usage{
\method{print}{signal}(x, ...)
}
\arguments{
\item{x}{A signal object.}

\item{...}{Additional arguments (not used).}
}
\description{
Provides a summary of the signal object.
}
