% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property.R
\name{property_relationships}
\alias{property_relationships}
\title{Get node pairs that are a certain number of path_lengths apart in an igraph object}
\usage{
property_relationships(path_length, relationships = NULL)
}
\arguments{
\item{path_length}{An integer indicating the number of path_lengths (edges) between nodes}

\item{relationships}{A vector of relationships, optional. If provided, it must match the path_length length.}
}
\value{
A data frame with pairs of nodes that are 'path_length' edges apart
}
\description{
Get node pairs that are a certain number of path_lengths apart in an igraph object
}
