% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_media.R
\name{rationalized_spatial_frequency_map}
\alias{rationalized_spatial_frequency_map}
\title{Rationalized Spatial Frequency Map}
\usage{
rationalized_spatial_frequency_map(spatial_frequencies)
}
\arguments{
\item{spatial_frequencies}{A matrix with each cell containing spatial frequencies (x, y).}
}
\value{
A matrix with rationalized spatial frequencies.
}
\description{
Creates a matrix where each cell contains a rationalized approximation of y / x using Stern-Brocot.
}
