% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{signal}
\alias{signal}
\title{Signal Constructor}
\usage{
signal(spectrum)
}
\arguments{
\item{spectrum}{An object of class "spectrum" (or one of its subclasses).}
}
\value{
An object of class \code{signal}.
}
\description{
Creates a signal object from a given spectrum object.
}
