% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_signal.R
\name{space_signal}
\alias{space_signal}
\title{Space Signal Constructor}
\usage{
space_signal(wavelength_spectrum)
}
\arguments{
\item{wavelength_spectrum}{An object of class "wavelength_spectrum" (subclass of "spectrum").}
}
\value{
An object of class \code{space_signal}.
}
\description{
Creates a space_signal object from a wavelength_spectrum object.
}
