% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_sbg.R
\name{sparse_spectrum_sbg}
\alias{sparse_spectrum_sbg}
\title{Sparse SBG kernel of max size}
\usage{
sparse_spectrum_sbg(kernel_size, uncertainty)
}
\arguments{
\item{kernel_size}{Size of the kernel (must be odd).}

\item{uncertainty}{Uncertainty product}
}
\value{
A 2D matrix representing the SBG kernel.
}
\description{
Generates a sparse SBG kernel with specific parameters.
}
