% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_signal.R
\name{time_signal}
\alias{time_signal}
\title{Time Signal Constructor}
\usage{
time_signal(frequency_spectrum)
}
\arguments{
\item{frequency_spectrum}{An object of class "frequency_spectrum" (subclass of "spectrum").}
}
\value{
An object of class \code{time_signal}.
}
\description{
Creates a time_signal object from a frequency_spectrum object.
}
