% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavyr-package.R
\docType{package}
\name{wavyr-package}
\alias{wavyr}
\alias{wavyr-package}
\title{wavyr: wavyr}
\description{
Allows researchers to explore composite waves in space and time.
}
\author{
\strong{Maintainer}: Brian Mulloy \email{brian@homeymusic.com} (\href{https://orcid.org/0000-0002-1803-3172}{ORCID})

}
\keyword{internal}
