% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_spectrum.R
\name{frequency_spectrum}
\alias{frequency_spectrum}
\title{Create a Frequency Spectrum}
\usage{
frequency_spectrum(frequency, amplitude = NULL, reference = NULL)
}
\arguments{
\item{frequency}{Either a numeric vector of frequencies or a list with named \code{frequency} and \code{amplitude} vectors.}

\item{amplitude}{A numeric vector of amplitudes, if \code{frequency} is a numeric vector.}

\item{reference}{Used to compute the fundamental frequency from the cycle length}
}
\value{
An object of class \code{frequency_spectrum} that inherits from \code{spectrum}.
}
\description{
Extends the \code{spectrum} class specifically for frequency spectra.
Accepts either direct numeric vectors or a list containing \code{frequency} and \code{amplitude}.
}
