% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_waveform.R
\name{linear_waveform}
\alias{linear_waveform}
\title{Create a LinearWaveform for a linear medium}
\usage{
linear_waveform(frequency_spectrum, speed_of_sound = 343)
}
\arguments{
\item{frequency_spectrum}{An object of class "frequency_spectrum" containing frequencies and amplitudes.}

\item{speed_of_sound}{Numeric, the speed of sound in the medium (e.g., 343 for air in m/s).}
}
\value{
An object of class "LinearWaveform" with frequency and wavelength spectra, and optional phase.
}
\description{
Create a LinearWaveform for a linear medium
}
