% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{plot.spectrum}
\alias{plot.spectrum}
\title{Plot a spectrum with spikes}
\usage{
\method{plot}{spectrum}(x, rectangles = numeric(0), ...)
}
\arguments{
\item{x}{An object of class "spectrum" containing components and amplitudes.}

\item{...}{Additional parameters for plotting.}
}
\description{
Creates a spike plot for a spectrum (frequency or wavelength).
}
