% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property.R
\name{property_relationships}
\alias{property_relationships}
\title{Get node pairs that are a specific path length apart in an igraph object}
\usage{
property_relationships(path_length, relationships = NULL)
}
\arguments{
\item{path_length}{An integer specifying the number of edges (hops) between nodes.}

\item{relationships}{A vector of relationships to filter edges by. If NULL, all relationships are considered.}
}
\value{
A data frame with pairs of nodes separated by the specified path length.
}
\description{
Get node pairs that are a specific path length apart in an igraph object
}
