% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{spectrum}
\alias{spectrum}
\alias{spectrum.default}
\title{Spectrum Constructor}
\usage{
spectrum(component, amplitude = NULL)

\method{spectrum}{default}(component, amplitude)
}
\arguments{
\item{component}{Either a numeric vector for component values or a list with named \code{component} and \code{amplitude} vectors.}

\item{amplitude}{A numeric vector of amplitudes, if \code{component} is a numeric vector.}
}
\value{
An object of class \code{spectrum}.
}
\description{
Creates a spectrum object with components and amplitudes.
Handles both direct vector inputs or a list containing named \code{component} and \code{amplitude}.
}
