% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{spectrum}
\alias{spectrum}
\alias{spectrum.default}
\title{Spectrum Constructor}
\usage{
spectrum(
  idealized_component,
  amplitude = NULL,
  extent_rate = EXTENT_RATE$rate,
  reference_component = NULL
)

\method{spectrum}{default}(
  idealized_component,
  amplitude,
  extent_rate = EXTENT_RATE$rate,
  reference_component = NULL
)
}
\arguments{
\item{idealized_component}{Either a numeric vector for component values or a list with named \code{idealized_component} and \code{amplitude} vectors.}

\item{amplitude}{A numeric vector of amplitudes, if \code{idealized_component} is a numeric vector.}

\item{extent_rate}{indictaes whether the wave represnetation is a rate (like frequency or wavenumber) or an extent (like period or wavelength)
If \code{FALSE}, treats the spectrum as a frequency domain (e.g., frequencies or wavenumbers).}

\item{reference_component}{For computing the fundamental component we take the product of the reference_component and the relative cycle length.}
}
\value{
An object of class \code{spectrum}.
}
\description{
Creates a spectrum object with components and amplitudes.
Handles both direct vector inputs or a list containing named \code{idealized_component} and \code{amplitude}.
}
