% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{spectrum}
\alias{spectrum}
\alias{spectrum.default}
\title{Spectrum Constructor}
\usage{
spectrum(
  component,
  amplitude = NULL,
  inverted = FALSE,
  reference_component = NULL
)

\method{spectrum}{default}(component, amplitude, inverted = FALSE, reference_component = NULL)
}
\arguments{
\item{component}{Either a numeric vector for component values or a list with named \code{component} and \code{amplitude} vectors.}

\item{amplitude}{A numeric vector of amplitudes, if \code{component} is a numeric vector.}

\item{inverted}{Logical; if \code{TRUE}, treats the spectrum as an inverted domain (e.g., periods or wavelengths).
If \code{FALSE}, treats the spectrum as a frequency domain (e.g., frequencies or wavenumbers).}

\item{reference_component}{For computing the fundamental component we take the product of the reference_component and the relative cycle length.}
}
\value{
An object of class \code{spectrum}.
}
\description{
Creates a spectrum object with components and amplitudes.
Handles both direct vector inputs or a list containing named \code{component} and \code{amplitude}.
}
