% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelength_spectrum.R
\name{wavelength_spectrum}
\alias{wavelength_spectrum}
\title{Create a wavelength spectrum}
\usage{
wavelength_spectrum(wavelengths)
}
\arguments{
\item{wavelengths}{A tibble with columns `wavelength` and `amplitude`}
}
\value{
An object of class "wavelength_spectrum" containing the wavelengths and amplitudes
}
\description{
Create a wavelength spectrum
}
